WI #2735 Refactor completion after qualified name separator
