+++ TypeCobol.LanguageServer/Completion Factory/CompletionFactory.cs
@@ -223,193 +223,6 @@ namespace TypeCobol.LanguageServer
 
         #endregion
 
-        #region QualifiedName Completion
-        public static List<CompletionItem> GetCompletionForQualifiedName(Position position, FileCompiler fileCompiler, CodeElement codeElement, Token qualifiedNameSeparatorToken, Token userFilterToken, Dictionary<SignatureInformation, FunctionDeclaration> functionDeclarationSignatureDictionary)
-        {
-            var completionItems = new List<CompletionItem>();
-            var arrangedCodeElement = codeElement as CodeElementWrapper;
-            var node = CompletionFactoryHelpers.GetMatchingNode(fileCompiler, codeElement);
-            if (node == null)
-                return completionItems;
-            var userFilterText = userFilterToken == null ? string.Empty : userFilterToken.Text;
-
-            //Get the token before MatchingToken 
-            var userTokenToSeek =
-                arrangedCodeElement?.ArrangedConsumedTokens.ElementAt(
-                    arrangedCodeElement.ArrangedConsumedTokens.IndexOf(qualifiedNameSeparatorToken) - 1);
-            var qualifiedNameTokens = new List<Token>();
-            if (arrangedCodeElement != null)
-            {
-                qualifiedNameTokens.AddRange(
-                    arrangedCodeElement.ArrangedConsumedTokens?.Where(
-                        t => (t?.TokenType == TokenType.UserDefinedWord || t?.TokenType == TokenType.QualifiedNameSeparator) && (t.EndColumn <= position.character && t.Line == position.line + 1) || t.Line < position.line + 1));
-                //Remove all the userdefinedword token and also QualifiedNameToken
-                arrangedCodeElement.ArrangedConsumedTokens = arrangedCodeElement.ArrangedConsumedTokens.Except(qualifiedNameTokens).ToList();
-                //We only wants the token that in front of any QualifiedName 
-                //Get the first significant token (i.e CALL/TYPE/...)
-                Token firstSignificantToken, tempUserFilterToken;
-                CodeElementMatcher.MatchCompletionCodeElement(position, new List<CodeElementWrapper> { arrangedCodeElement }, out tempUserFilterToken,
-                    out firstSignificantToken);
-
-
-                //Select the qualifiedName chain closest to cursor
-                Token previousToken = null;
-                qualifiedNameTokens.Reverse();
-                var filteredQualifiedNameTokens = new List<Token>(); //Will contains all the tokens forming the qualifiedName chain. 
-                foreach (var token in qualifiedNameTokens)
-                {
-                    if (previousToken == null 
-                        || ((previousToken.TokenType == TokenType.QualifiedNameSeparator && token.TokenType == TokenType.UserDefinedWord) 
-                            || (token.TokenType == TokenType.QualifiedNameSeparator && previousToken.TokenType == TokenType.UserDefinedWord)))
-                        filteredQualifiedNameTokens.Add(token);
-                    else
-                        break;
-
-                    previousToken = token;
-                }
-                filteredQualifiedNameTokens.Reverse();
-
-               //For MOVE INPUT OUTPUT variables etc.. , get all the children of a variable that are accessible
-               //Try to find corresponding variables
-               var qualifiedName = 
-                    filteredQualifiedNameTokens.Where(
-                            t =>
-                                t.TokenType == TokenType.UserDefinedWord &&
-                                !(t.Text == userFilterText && userFilterToken != null &&
-                                  t.StartIndex == userFilterToken.StartIndex && t.EndColumn == userFilterToken.EndColumn) &&
-                                ((firstSignificantToken != null && ((t.StartIndex >= firstSignificantToken.EndColumn && t.Line == firstSignificantToken.Line) || t.Line > firstSignificantToken.Line)) 
-                                || firstSignificantToken == null) 
-                                && ((t.EndColumn <= position.character && t.Line == position.line + 1) || t.Line < position.line + 1))
-                        .Select(t => t.Text)
-                        .ToArray();
-
-               var possibleVariables = qualifiedName.Length > 0 ? node.SymbolTable.GetVariablesExplicit(new URI(qualifiedName)).ToArray() : null;
-
-                List<DataDefinition> childrenCandidates = new List<DataDefinition>();
-                if (possibleVariables != null && possibleVariables.Length > 0) 
-                {
-                    //Get children of a type to get completion possibilities
-                    foreach (var variable in possibleVariables)
-                    {
-                        var children = new List<Node>();
-
-                        //if it's a typed variable, propose 1st children of the type
-                        if (variable.TypeDefinition != null)
-                        {
-                            children.AddRange(variable.TypeDefinition.Children);
-                        }
-
-                        //It's a variable with levels inside
-                        //Note that Index are also children
-                        if (variable.Children?.Count > 0)
-                            children.AddRange(variable.Children);
-                        //No children => nothing to do
-
-                        var computedChildrenList = new List<Node>();
-                        foreach (var child in children)
-                        {
-                            GetNextRelevantChildren(child, computedChildrenList);
-                        }
-
-                        FilterByUserText(computedChildrenList);
-                    }
-                }
-                else
-                {
-                    //If no variable found, it could be a child declared in a typedef.
-                    var children = new List<Node>();
-                    var potentialTypes =
-                        node.SymbolTable.GetTypes(
-                            t =>
-                                t.Children != null &&
-                                t.Children.Any(
-                                    tc => tc.Name != null && tc.Name.Equals(userTokenToSeek.Text, StringComparison.InvariantCultureIgnoreCase)),
-                                SymbolTable.Scope.Namespace
-                            );
-
-                    foreach (var nodeType in potentialTypes.SelectMany(t => t.Children).Where(c => c != null && c.Name != null && c.Name.Equals(userTokenToSeek.Text, StringComparison.InvariantCultureIgnoreCase)))
-                    {
-                        var nodeDataDef = nodeType as DataDefinition;
-                        if (nodeDataDef == null) continue;
-
-                        var typeChildren = GetTypeChildren(node.SymbolTable, nodeDataDef);
-                        if (typeChildren != null)
-                            children.AddRange(typeChildren);
-                    }
-
-                    FilterByUserText(children);
-                }
-
-                void FilterByUserText(List<Node> nodes)
-                {
-                    foreach (var child in nodes)
-                    {
-                        System.Diagnostics.Debug.Assert(child is DataDefinition);
-                        var data = (DataDefinition) child;
-                        if (data.Name != null && data.Name.StartsWith(userFilterText, StringComparison.InvariantCultureIgnoreCase))
-                        {
-                            childrenCandidates.Add(data);
-                        }
-                    }
-                }
-
-                var variables = childrenCandidates.Select(v => new KeyValuePair<DataDefinitionPath, DataDefinition>(null, v));
-                var items = CompletionFactoryHelpers.CreateCompletionItemsForVariableSetAndDisambiguate(variables, fileCompiler.CompilerOptions, true);
-                completionItems.AddRange(items);
-
-                if (firstSignificantToken != null)
-                {
-                    switch (firstSignificantToken.TokenType)
-                    {
-                        case TokenType.CALL:
-                        {
-                            functionDeclarationSignatureDictionary.Clear(); //Clear to avoid key collision
-                            //On CALL get possible procedures and functions in the seeked program
-                            var program = node.SymbolTable.GetPrograms(userTokenToSeek.Text, true).FirstOrDefault();
-                            if (program != null)
-                            {
-                                var procedures =
-                                    program
-                                        .SymbolTable.GetFunctions(
-                                            f =>
-                                                f.Name.StartsWith(userFilterText,
-                                                    StringComparison.InvariantCultureIgnoreCase) ||
-                                                f.VisualQualifiedName.ToString()
-                                                    .StartsWith(userFilterText,
-                                                        StringComparison.InvariantCultureIgnoreCase),
-                                            SymbolTable.Scope.Program);
-
-                                completionItems.AddRange(CompletionFactoryHelpers.CreateCompletionItemsForProcedures(procedures, node, functionDeclarationSignatureDictionary, false));
-
-                            }
-                            break;
-                        }
-                        case TokenType.TYPE:
-                        {
-                            //On TYPE get possible public types in the seeked program
-                            var program = node.SymbolTable.GetPrograms(userTokenToSeek.Text, true).FirstOrDefault();
-                            if (program != null)
-                            {
-                                var types =
-                                    program
-                                        .SymbolTable.GetTypes(
-                                            t =>
-                                                t.Name.StartsWith(userFilterText,
-                                                    StringComparison.InvariantCultureIgnoreCase),
-                                                SymbolTable.Scope.Program
-                                            );
-                                completionItems.AddRange(CompletionFactoryHelpers.CreateCompletionItemsForType(types, node, false));
-                            }
-                            break;
-                        }
-                    }
-                }
-            }
-
-            return completionItems.Distinct().ToList();
-        }
-        #endregion
-
         #region Variable Completion
         public static List<CompletionItem> GetCompletionForVariable(FileCompiler fileCompiler, CodeElement codeElement, Func<DataDefinition, bool> predicate)
         {
@@ -725,40 +538,6 @@ namespace TypeCobol.LanguageServer
                 completionItems.Add(CompletionFactoryHelpers.CreateCompletionItemForSingleVariable(null, currentParent, options, true));
             }
             return completionItems;
-        }
-
-        #endregion
-
-        #region Helpers
-
-        private static IReadOnlyList<Node> GetTypeChildren(SymbolTable symbolTable, DataDefinition dataDefNode)
-        {
-            if (symbolTable == null || dataDefNode == null)
-                return null;
-
-            var type = symbolTable.GetTypes(
-                t => t.Name.Equals(dataDefNode.DataType.Name, StringComparison.InvariantCultureIgnoreCase)
-                     ||
-                     t.VisualQualifiedName.ToString()
-                         .Equals(dataDefNode.DataType.Name, StringComparison.InvariantCultureIgnoreCase),
-                    SymbolTable.Scope.Intrinsic).FirstOrDefault();
-
-            return type?.Children;
-        }
-
-        private static void GetNextRelevantChildren(Node dataDefinition, List<Node> children)
-        {
-            if (string.IsNullOrEmpty(dataDefinition.Name) && dataDefinition.Children.Any())
-            {
-                foreach (var child in dataDefinition.Children)
-                {
-                    GetNextRelevantChildren(child, children);
-                }
-            }
-            else
-            {
-                children.Add(dataDefinition);
-            }
         }
 
         #endregion
